added FLGAN
First, I implemented FLGAN according to GENERATIVE MODELS FOR EFFECTIVE ML ON PRIVATE, DECENTRALIZED DATASETS; however, the performance was not good. Therefore, instead of separately training DNet and GNet in an epoch for a local participant, I trained DNet and GNet together within an epoch for a local user. Still, the performance was not acceptable. I will try to improve the aggregation. //TODO